+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Created `ksd` documentation. #19
+- Added a `SheetParserFactory.Feature.DISABLE_VULNERABILITY_PROTECTIONS` to disable vulnerability detection
+  such as Zip bombs. #33
+- Added a new option `KeyedSheetDiff` to disable disable vulnerability detection. #33
 
 ### Changed
 - Improved error message in Header.